+++ docs/constants/navigation.js
@@ -117,12 +117,16 @@ const home = [
     makePage('develop/authentication.mdx'),
     makePage('develop/unit-testing.mdx'),
   ]),
-  makeSection('Review', [makePage('review/overview.mdx'), makePage('review/with-orbit.mdx')]),
+  makeSection('Review', [
+    makePage('review/overview.mdx'),
+    makePage('review/share-previews-with-your-team.mdx'),
+    makePage('review/with-orbit.mdx'),
+  ]),
   makeSection('Deploy', [
     makePage('deploy/build-project.mdx'),
     makePage('deploy/submit-to-app-stores.mdx'),
     makePage('deploy/app-stores-metadata.mdx'),
-    makePage('deploy/instant-updates.mdx'),
+    makePage('deploy/send-over-the-air-updates.mdx'),
   ]),
   makeSection('More', [makePage('core-concepts.mdx'), makePage('faq.mdx')]),
 ];
@@ -200,7 +204,6 @@ const general = [
     makePage('router/create-pages.mdx'),
     makePage('router/navigating-pages.mdx'),
     makePage('router/layouts.mdx'),
-    makePage('router/appearance.mdx'),
     makePage('router/error-handling.mdx'),
     makeGroup('Navigation patterns', [
       makePage('router/advanced/root-layout.mdx'),
@@ -284,6 +287,7 @@ const general = [
         makePage('app-signing/existing-credentials.mdx'),
         makePage('app-signing/syncing-credentials.mdx'),
         makePage('app-signing/security.mdx'),
+        makePage('app-signing/apple-developer-program-roles-and-permissions.mdx'),
       ],
       { expanded: false }
     ),
@@ -361,7 +365,6 @@ const general = [
     ]),
     makeGroup('Reference', [
       makePage('eas-update/migrate-from-classic-updates.mdx'),
-      makePage('eas-update/custom-updates-server.mdx'),
       makePage('eas-update/codepush.mdx'),
       makePage('eas-update/updating-your-app.mdx'),
     ]),
@@ -410,7 +413,6 @@ const general = [
       makePage('eas/webhooks.mdx'),
       makeSection('Assorted', [
         makePage('guides/authentication.mdx'),
-        makePage('guides/troubleshooting-proxies.mdx'),
         makePage('guides/sharing-preview-releases.mdx'),
         makePage('guides/using-hermes.mdx'),
         makePage('guides/ios-developer-mode.mdx'),
@@ -420,6 +422,8 @@ const general = [
         makePage('guides/using-bun.mdx'),
         makePage('guides/editing-richtext.mdx'),
         makePage('guides/store-assets.mdx'),
+        makePage('guides/local-first.mdx'),
+        makePage('guides/keyboard-handling.mdx'),
       ]),
       makeSection('Integrations', [
         makePage('guides/using-analytics.mdx'),
@@ -448,6 +452,14 @@ const general = [
         makePage('billing/invoices-and-receipts.mdx'),
         makePage('billing/usage-based-pricing.mdx'),
         makePage('billing/faq.mdx'),
+      ]),
+      makeSection('Troubleshooting', [
+        makePage('troubleshooting/overview.mdx'),
+        makePage('troubleshooting/application-has-not-been-registered.mdx'),
+        makePage('troubleshooting/clear-cache-macos-linux.mdx'),
+        makePage('troubleshooting/clear-cache-windows.mdx'),
+        makePage('troubleshooting/react-native-version-mismatch.mdx'),
+        makePage('troubleshooting/proxies.mdx'),
       ]),
     ],
     { expanded: true }
@@ -552,6 +564,7 @@ const archive = [
     makePage('archive/customizing-webpack.mdx'),
     makePage('archive/using-expo-client.mdx'),
     makePage('archive/using-flipper.mdx'),
+    makePage('archive/e2e-tests.mdx'),
     makePage('archive/glossary.mdx'),
   ]),
 ];