+++ packages/expo-audio/src/AudioModule.types.ts
@@ -28,7 +28,7 @@ export declare class AudioPlayer extends SharedObject<AudioEvents> {
    * Initializes a new audio player instance with the given source.
    * @hidden
    */
-  constructor(source: AudioSource | string | number | null);
+  constructor(source: AudioSource, updateInterval: number);
 
   /**
    * Unique identifier for the player object.
@@ -61,6 +61,11 @@ export declare class AudioPlayer extends SharedObject<AudioEvents> {
   isLoaded: boolean;
 
   /**
+   * Boolean value indicating whether audio sampling is supported on the platform.
+   */
+  isAudioSamplingSupported: boolean;
+
+  /**
    * Boolean value indicating whether the player is buffering.
    */
   isBuffering: boolean;
@@ -91,6 +96,12 @@ export declare class AudioPlayer extends SharedObject<AudioEvents> {
   shouldCorrectPitch: boolean;
 
   /**
+   * The current status of the audio player.
+   * @hidden
+   */
+  currentStatus: AudioStatus;
+
+  /**
    * Resumes the player.
    */
   play(): void;
@@ -114,13 +125,25 @@ export declare class AudioPlayer extends SharedObject<AudioEvents> {
   setPlaybackRate(second: number, pitchCorrectionQuality?: PitchCorrectionQuality): void;
 
   /**
-   * Release the player and frees up resources.
+   *
+   * @hidden
    */
-  release(): void;
+  setAudioSamplingEnabled(enabled: boolean): void;
+
+  /**
+   * Remove the player from memory to free up resources.
+   */
+  remove(): void;
 }
 
-type AudioEvents = {
+type AudioSample = {
+  channels: { frames: number[] }[];
+  timestamp: number;
+};
+
+export type AudioEvents = {
   onPlaybackStatusUpdate(status: AudioStatus): void;
+  onAudioSampleUpdate(data: AudioSample): void;
 };
 
 export declare class AudioRecorder extends SharedObject<RecordingEvents> {
@@ -196,15 +219,15 @@ export declare class AudioRecorder extends SharedObject<RecordingEvents> {
   startRecordingAtTime(seconds: number): void;
 
   /**
+   * Prepares the recording for recording.
+   */
+  prepareToRecordAsync(): Promise<void>;
+
+  /**
    * Stops the recording once the specified time has elapsed.
    * @param seconds The time in seconds to stop recording at.
    */
   recordForDuration(seconds: number): void;
-
-  /**
-   * Release the recorder and frees up resources.
-   */
-  release(): void;
 }
 
 export type RecordingEvents = {